refactor(hooks): extract SubagentStop no-commit gate from hook_router (#35, #2384)
Wave-2 PR1 of the hook_router split. Pure behavior-preserving move of
_record_no_commit_signal + handle_subagent_stop_no_commit out of the
hook_router god-module into a new flat sibling
hooks/scripts/subagent_no_commit.py, re-exported into the router import
block so the SubagentStop dispatch entry and every existing test stay
unchanged.

The sibling follows the Wave-1 extraction pattern: dual-identity sys.modules
alias, module-top imports limited to stdlib + the already-extracted
state_files sibling, and a lazy back-import of the spine helpers
(_ensure_state_dir, _state_file) inside the handler body. The pure decision
leaf teatree.hooks.no_commit_detector is unchanged and still imported
function-scoped after the src bootstrap.

Module-health: hook_router shrinks (code-LOC 5814 -> 5761, public fns
55 -> 54); the new sibling is 86 code-LOC / 1 public fn. The intra-core
deferred-import ratchet is unchanged (this move is outside src/teatree/core).
The only test edit is the single mechanical repoint of the
_record_no_commit_signal monkeypatch target to the new module, plus a
cold-import smoke assertion.

Refs #2384.
